+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - "Play" context menu item in playlists view
 - "Jump to current track" context menu item status bar
 - "Keyboard shortcuts" menu item
+- Ability to sort library tree by date
 
 ### Misc
 - Update Taglib to 1.12